requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0; python_version >= '3.7'
backports.zoneinfo==0.2.1; python_version < '3.9'
django==4.0.3
feedparser==6.0.8
python-dateutil==2.8.2
python-dotenv==0.19.2
sgmllib3k==1.0.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
tzdata==2021.5; sys_platform == 'win32'
